+++ R/Spectra.R
@@ -1519,6 +1519,7 @@ setMethod(
              BPPARAM = bpparam()) {
         backend_class <- class(object@backend)
         BPPARAM <- backendBpparam(object@backend, BPPARAM)
+        BPPARAM <- backendBpparam(backend, BPPARAM)
         if (!supportsSetBackend(backend))
             stop(class(backend), " does not support 'setBackend'")
         if (!length(object)) {
@@ -1528,7 +1529,7 @@ setMethod(
             f <- force(factor(f, levels = unique(f)))
             if (length(f) != length(object))
                 stop("length of 'f' has to match the length of 'object'")
-            if (length(levels(f)) == 1L) {
+            if (length(levels(f)) == 1L || inherits(BPPARAM, "SerialParam")) {
                 bknds <- backendInitialize(
                     backend, data = spectraData(object@backend), ...)
             } else {